requirements.txt
africastalking==1.2.7
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
django-oauth-toolkit==2.4.0
django-oidc-provider==0.8.2
django-phonenumber-field==7.3.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-social-oauth2==2.2.0
drf-spectacular==0.27.2
exceptiongroup==1.2.1
factory-boy==3.3.0
fake-factory==9999.9.9
Faker==18.9.0
filelock==3.15.3
future==1.0.0
identify==2.5.36
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
josepy==1.14.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
Markdown==3.6
mozilla-django-oidc==4.0.1
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.1
phonenumbers==8.13.39
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
psycopg2==2.9.9
pycparser==2.22
pycryptodomex==3.20.0
pyjwkest==1.4.2
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytest==8.2.2
pytest-django==4.8.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
schema==0.7.7
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.2
